Analysis of the volume of tweets from different regions
This analysis was based on the augmented database obtained by fetching tweets from the timelines of the users already present in the initial database. Here, the initial database refers to all the tweets that were obtained 2 weeks before or and after the earthquake from users tweeting with their location in Nepal.
The augmented database represents additional tweets from the same users which didn't have location on and hence were not captured in the initial database.